Show newsletter popup after 10s with embedded signup form
Extract the Mailchimp form from the on-page article newsletter block
into a shared NewsletterForm component, and use it inside the blog
popup instead of linking out to the Mailchimp landing page.

Also drops the popup delay from 15s to 10s.
